requirements.txt
kafka-python
Faker
pyodbc~=5.1.0
pandas~=2.2.2